+++ common/src/main/java/es/urjc/etsii/grafo/algorithms/SimpleAlgorithm.java
@@ -5,9 +5,10 @@ import es.urjc.etsii.grafo.annotations.ProvidedParam;
 import es.urjc.etsii.grafo.create.Constructive;
 import es.urjc.etsii.grafo.improve.Improver;
 import es.urjc.etsii.grafo.io.Instance;
+import es.urjc.etsii.grafo.metrics.BestObjective;
 import es.urjc.etsii.grafo.solution.Solution;
-import es.urjc.etsii.grafo.solution.metrics.Metrics;
-import es.urjc.etsii.grafo.solution.metrics.MetricsManager;
+import es.urjc.etsii.grafo.metrics.Metrics;
+import es.urjc.etsii.grafo.metrics.MetricsManager;
 import es.urjc.etsii.grafo.util.StringUtil;
 import es.urjc.etsii.grafo.util.TimeControl;
 import es.urjc.etsii.grafo.util.ValidationUtil;
@@ -67,7 +68,7 @@ public class SimpleAlgorithm<S extends Solution<S, I>, I extends Instance> exten
         var solution = this.newSolution(instance);
         solution = constructive.construct(solution);
         ValidationUtil.assertValidScore(solution);
-        MetricsManager.addDatapoint(Metrics.BEST_OBJECTIVE_FUNCTION, solution.getScore());
+        BestObjective.add(solution.getScore());
         printStatus("Constructive", solution);
         solution = localSearch(solution);
         return solution;
@@ -85,7 +86,7 @@ public class SimpleAlgorithm<S extends Solution<S, I>, I extends Instance> exten
         }
         solution = improver.improve(solution);
         ValidationUtil.assertValidScore(solution);
-        MetricsManager.addDatapoint(Metrics.BEST_OBJECTIVE_FUNCTION, solution.getScore());
+        BestObjective.add(solution.getScore());
         printStatus("Improver " + improver, solution);
         return solution;
     }