+++ common/src/main/java/es/urjc/etsii/grafo/improve/Improver.java
@@ -5,9 +5,10 @@ import es.urjc.etsii.grafo.annotations.AlgorithmComponent;
 import es.urjc.etsii.grafo.annotations.AutoconfigConstructor;
 import es.urjc.etsii.grafo.annotations.ProvidedParam;
 import es.urjc.etsii.grafo.io.Instance;
+import es.urjc.etsii.grafo.metrics.BestObjective;
 import es.urjc.etsii.grafo.solution.Solution;
-import es.urjc.etsii.grafo.solution.metrics.Metrics;
-import es.urjc.etsii.grafo.solution.metrics.MetricsManager;
+import es.urjc.etsii.grafo.metrics.Metrics;
+import es.urjc.etsii.grafo.metrics.MetricsManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -58,7 +59,7 @@ public abstract class Improver<S extends Solution<S,I>,I extends Instance> {
             throw new IllegalStateException(String.format("Score has worsened after executing an improvement method: %s --> %s", initialScore, endScore));
         }
         if(ofmode.isBetter(endScore, initialScore)){
-            MetricsManager.addDatapoint(Metrics.BEST_OBJECTIVE_FUNCTION, endScore);
+            BestObjective.add(endScore);
         }
         return improvedSolution;
     }