+++ common/src/test/java/es/urjc/etsii/grafo/metrics/MetricsTest.java
@@ -1,4 +1,4 @@
-package es.urjc.etsii.grafo.solution.metrics;
+package es.urjc.etsii.grafo.metrics;
 
 import org.junit.jupiter.api.Test;
 
@@ -12,12 +12,12 @@ class MetricsTest {
     void basicTest() {
         long currentTime = System.nanoTime();
         var metric = new Metrics();
-        long diff = metric.referencePoint - currentTime;
+        long diff = metric.referenceTime - currentTime;
         assertTrue(diff < ERROR_MARGIN); // should use nanoTime, with a margin for error of 50ms
 
         currentTime = System.nanoTime();
         metric.addDatapoint("test", 76);
-        assertThrows(IllegalArgumentException.class, () -> metric.addDatapoint("any", 10, 10));
+        assertThrows(IllegalArgumentException.class, () -> metric.addDatapoint("any", initializeMetric, 10, 10));
         assertThrows(IllegalArgumentException.class, () -> metric.byName("any"));
         var set = metric.byName("test");
         assertEquals(1, set.size());
@@ -31,27 +31,27 @@ class MetricsTest {
     @Test
     void mergeAndContains() {
         var metricA = new Metrics(10);
-        metricA.addDatapoint("testA", 1, 31);
-        metricA.addDatapoint("testA", 5, 35);
+        metricA.addDatapoint("testA", initializeMetric, 1, 31);
+        metricA.addDatapoint("testA", initializeMetric, 5, 35);
         var metricB = new Metrics(20);
-        metricB.addDatapoint("testA", 4, 34);
-        metricB.addDatapoint("testB", 8, 38);
+        metricB.addDatapoint("testA", initializeMetric, 4, 34);
+        metricB.addDatapoint("testB", initializeMetric, 8, 38);
 
         var metricC = new Metrics(30);
-        metricC.addDatapoint("testB", 6, 36);
+        metricC.addDatapoint("testB", initializeMetric, 6, 36);
 
 
         var merged = Metrics.merge(metricA, metricB, metricC);
-        assertEquals(10, merged.referencePoint); // Lowest reference point of all metrics
-        assertEquals(2, merged.metricValues.size());
-        assertEquals(3, merged.metricValues.get("testA").size());
-        assertEquals(2, merged.metricValues.get("testB").size());
+        assertEquals(10, merged.referenceTime); // Lowest reference point of all metrics
+        assertEquals(2, merged.metrics.size());
+        assertEquals(3, merged.metrics.get("testA").size());
+        assertEquals(2, merged.metrics.get("testB").size());
 
-        var orderedA = merged.metricValues.get("testA").toArray(TYPEREF);
+        var orderedA = merged.metrics.get("testA").toArray(TYPEREF);
         assertEquals(1, orderedA[0].value());
         assertEquals(4, orderedA[1].value());
         assertEquals(5, orderedA[2].value());
-        var orderedB = merged.metricValues.get("testB").toArray(TYPEREF);
+        var orderedB = merged.metrics.get("testB").toArray(TYPEREF);
         assertEquals(6, orderedB[0].value());
         assertEquals(8, orderedB[1].value());
 
@@ -62,10 +62,10 @@ class MetricsTest {
         var metricA = new Metrics(10);
         // calculate area between 15 and 35, or in other words, ignore first 5 nanoseconds, duration 30
         assertThrows(IllegalArgumentException.class, () -> metricA.areaUnderCurve("test", 5, 30));
-        assertDoesNotThrow(() -> metricA.addDatapoint("test", 5, 11));
-        assertDoesNotThrow(() -> metricA.addDatapoint("test", 2, 21));
-        assertDoesNotThrow(() -> metricA.addDatapoint("test", 1, 31));
-        assertDoesNotThrow(() -> metricA.addDatapoint("test", 0.5, 41));
+        assertDoesNotThrow(() -> metricA.addDatapoint("test", initializeMetric, 5, 11));
+        assertDoesNotThrow(() -> metricA.addDatapoint("test", initializeMetric, 2, 21));
+        assertDoesNotThrow(() -> metricA.addDatapoint("test", initializeMetric, 1, 31));
+        assertDoesNotThrow(() -> metricA.addDatapoint("test", initializeMetric, 0.5, 41));
         assertThrows(ArithmeticException.class, () -> metricA.areaUnderCurve("test", 5, Long.MAX_VALUE));
         assertThrows(IllegalArgumentException.class, () -> metricA.areaUnderCurve("test", 0, 30));
         double area1 = metricA.areaUnderCurve("test", 1, 35);