+++ core/src/main/java/es/urjc/etsii/grafo/solver/Mork.java
@@ -2,6 +2,7 @@ package es.urjc.etsii.grafo.solver;
 
 import es.urjc.etsii.grafo.algorithms.FMode;
 import es.urjc.etsii.grafo.annotations.InheritedComponent;
+import es.urjc.etsii.grafo.metrics.MetricsManager;
 import es.urjc.etsii.grafo.services.BannerProvider;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -67,6 +68,7 @@ public class Mork {
      */
     public static void setSolvingMode(FMode fmode) {
         Mork.fmode = fmode;
+        MetricsManager.setSolvingMode(fmode);
     }
 
     /**